feat: separated into namespaces
