+++ ws-service/types.ts
@@ -1,0 +1,8 @@
+import { TransactionWithAddressAndPrices } from 'services/transactionService'
+
+type TxBroadcast = 'NewTx' | 'OldTx'
+export interface BroadcastTxData {
+  address: string
+  txs: TransactionWithAddressAndPrices[]
+  messageType: TxBroadcast
+}
